refactor(account-utils): file changed asnd used in test file for offchain and zkp microservice
