+++ zkp/__tests__/f-token-controller.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-unresolved */
 
 import Utils from 'zkp-utils';
-import AccountUtils from '../src/account-utils/account-utils';
+import AccountUtils from '../src/account-utils';
 
 import controller from '../src/f-token-controller';
